// src/common/enums/post.enums.ts
//media type
export enum MediaTypeEnum {
  IMAGE = "Image",
  VIDEO = "Video",
}

//visibility
export enum PostVisibilityEnum {
  PUBLIC = "Public",
  PRIVATE = "Private",
  FOLLOWERS_ONLY = "Followers_Only",
}

//status
export enum PostStatusEnum {
  PROCESSING = "Processing",
  PUBLISHED = "Published",
  FAILED = "Failed",
}