// src/common/enums/profile.enums.ts
//visibility
export enum ProfileVisibilityEnum {
  PUBLIC = "Public",
  PRIVATE = "Private",
}

//language
export enum LanguageEnum {
  ARABIC = "Arabic",
  ENGLISH = "English",
}

//gender
export enum GenderEnum {
  MALE = "Male",
  FEMALE = "Female",
}

//relation
export enum RelationEnum {
  SINGLE = "Single",
  MARRIED = "Married",
}